Swap '0' at the end of the Array.
Given an integer array 'nums', move all 0's to the end of it while maintaining the relative order of the non-zero elements.

Note: that you must do this in-place without making a copy of the array.

For that we can use swapping method. Every '0' element would be swapped  with every non- zero element.